+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 <a name="breaking_changes_1.30.0">[Breaking Changes:](#breaking_changes_1.30.0)</a>
 
 - [core] Added constructor injection to `ApplicationShell`: `SecondaryWindowHandler`. [#11048](https://github.com/eclipse-theia/theia/pull/11048) - Contributed on behalf of ST Microelectronics and Ericsson and by ARM and EclipseSource
+- [core] Changed type of `FrontendApplicationConfig#defaultTheme` from `string` to `DefaultTheme`. From now on, the default theme can be dispatched based on the OS theme. Use `DefaultTheme#defaultForOSTheme` to derive the `string` theme ID.
 
 ## v1.29.0 - 8/25/2022